Turn Queue can now operate.
The battle will start and wait for each combatant to make a move until moving on to the next one. Changed Character class to be abstract to allow for different behavior in entities (player v npc). Added Faction and Move scriptable objects to allow for multiple moves and groups to be made with different funcitonality.