Queue Rework Cont. and Battle Changes
Finished the general flow of the Queue Rework to make methods less dependent on state when they shouldn't be. Battle also now allows for joining into fray and leaving battle, and requires at least 2 combatants and opposing forces before actually starting the battle. Also changes made to the 32f branch has been translated to this branch as well for the most part.